Refactor Locations module test suite and mocks
- Rewrote and reorganized unit tests for all application handlers with clearer structure and improved coverage
- Split API mapper extension tests into dedicated files
- Removed redundant usings and clarified comments (some localized)
- Extracted and enhanced MockHttpMessageHandler for shared use across HTTP client tests
- Updated test project to include shared mock infrastructure
- Improved assertion messages and test intent for better maintainability and readability
