+++ src/Modules/Locations/Tests/Integration/Infrastructure/Queries/DbContextAllowedCityQueriesTests.cs
@@ -38,6 +38,7 @@ public class DbContextAllowedCityQueriesTests : BaseDatabaseTest
     [Fact]
     public async Task GetAllActiveAsync_ShouldReturnOnlyActiveCities()
     {
+        // Arrange
         var activeCity1 = AllowedCityBuilder.AsTestCity("Muriaé", "MG").Build();
         var activeCity2 = AllowedCityBuilder.AsTestCity("Itaperuna", "RJ").Build();
         var inactiveCity = AllowedCityBuilder.AsTestCity("São Paulo", "SP").AsInactive().Build();
@@ -47,8 +48,10 @@ public class DbContextAllowedCityQueriesTests : BaseDatabaseTest
         _context.AllowedCities.Add(inactiveCity);
         await _context.SaveChangesAsync();
 
+        // Act
         var result = await _queries.GetAllActiveAsync();
 
+        // Assert
         result.Should().HaveCount(2);
         result.Should().AllSatisfy(c => c.IsActive.Should().BeTrue());
         result.Should().Contain(c => c.CityName == "Muriaé");
@@ -59,6 +62,7 @@ public class DbContextAllowedCityQueriesTests : BaseDatabaseTest
     [Fact]
     public async Task GetAllAsync_ShouldReturnAllCities()
     {
+        // Arrange
         var activeCity = AllowedCityBuilder.AsTestCity("Muriaé", "MG").Build();
         var inactiveCity = AllowedCityBuilder.AsTestCity("Itaperuna", "RJ").AsInactive().Build();
 
@@ -66,8 +70,10 @@ public class DbContextAllowedCityQueriesTests : BaseDatabaseTest
         _context.AllowedCities.Add(inactiveCity);
         await _context.SaveChangesAsync();
 
+        // Act
         var result = await _queries.GetAllAsync();
 
+        // Assert
         result.Should().HaveCount(2);
         result.Should().Contain(c => c.IsActive);
         result.Should().Contain(c => !c.IsActive);
@@ -76,12 +82,15 @@ public class DbContextAllowedCityQueriesTests : BaseDatabaseTest
     [Fact]
     public async Task GetByCityAndStateAsync_WithExistingCityAndState_ShouldReturnCity()
     {
+        // Arrange
         var city = AllowedCityBuilder.AsTestCity("Muriaé", "MG").Build();
         _context.AllowedCities.Add(city);
         await _context.SaveChangesAsync();
 
+        // Act
         var result = await _queries.GetByCityAndStateAsync("Muriaé", "MG");
 
+        // Assert
         result.Should().NotBeNull();
         result!.CityName.Should().Be("Muriaé");
         result.StateSigla.Should().Be("MG");
@@ -90,66 +99,88 @@ public class DbContextAllowedCityQueriesTests : BaseDatabaseTest
     [Fact]
     public async Task GetByCityAndStateAsync_WithNonExistingCityAndState_ShouldReturnNull()
     {
+        // Arrange
+
+        // Act
         var result = await _queries.GetByCityAndStateAsync("Não Existe", "XX");
 
+        // Assert
         result.Should().BeNull();
     }
 
     [Fact]
     public async Task IsCityAllowedAsync_WithActiveCity_ShouldReturnTrue()
     {
+        // Arrange
         var city = AllowedCityBuilder.AsTestCity("Muriaé", "MG").Build();
         _context.AllowedCities.Add(city);
         await _context.SaveChangesAsync();
 
+        // Act
         var result = await _queries.IsCityAllowedAsync("Muriaé", "MG");
 
+        // Assert
         result.Should().BeTrue();
     }
 
     [Fact]
     public async Task IsCityAllowedAsync_WithInactiveCity_ShouldReturnFalse()
     {
+        // Arrange
         var city = AllowedCityBuilder.AsTestCity("Muriaé", "MG").AsInactive().Build();
         _context.AllowedCities.Add(city);
         await _context.SaveChangesAsync();
 
+        // Act
         var result = await _queries.IsCityAllowedAsync("Muriaé", "MG");
 
+        // Assert
         result.Should().BeFalse();
     }
 
     [Fact]
     public async Task IsCityAllowedAsync_WithNonExistingCity_ShouldReturnFalse()
     {
+        // Arrange
+
+        // Act
         var result = await _queries.IsCityAllowedAsync("Não Existe", "XX");
 
+        // Assert
         result.Should().BeFalse();
     }
 
     [Fact]
     public async Task ExistsAsync_WithExistingCity_ShouldReturnTrue()
     {
+        // Arrange
         var city = AllowedCityBuilder.AsTestCity("Muriaé", "MG").Build();
         _context.AllowedCities.Add(city);
         await _context.SaveChangesAsync();
 
+        // Act
         var result = await _queries.ExistsAsync("Muriaé", "MG");
 
+        // Assert
         result.Should().BeTrue();
     }
 
     [Fact]
     public async Task ExistsAsync_WithNonExistingCity_ShouldReturnFalse()
     {
+        // Arrange
+
+        // Act
         var result = await _queries.ExistsAsync("Não Existe", "XX");
 
+        // Assert
         result.Should().BeFalse();
     }
 
     [Fact]
     public async Task GetAllActiveAsync_ShouldReturnOrderedByStateAndCityName()
     {
+        // Arrange
         var city1 = AllowedCityBuilder.AsTestCity("Muriaé", "MG").WithIbgeCode(3143906).Build();
         var city2 = AllowedCityBuilder.AsTestCity("Itaperuna", "RJ").WithIbgeCode(3302270).Build();
         var city3 = AllowedCityBuilder.AsTestCity("Bom Jesus do Itabapoana", "RJ").WithIbgeCode(3300704).Build();
@@ -159,8 +190,10 @@ public class DbContextAllowedCityQueriesTests : BaseDatabaseTest
         _context.AllowedCities.Add(city3);
         await _context.SaveChangesAsync();
 
+        // Act
         var result = await _queries.GetAllActiveAsync();
 
+        // Assert
         result.Should().HaveCount(3);
         result[0].StateSigla.Should().Be("MG");
         result[0].CityName.Should().Be("Muriaé");
@@ -173,13 +206,16 @@ public class DbContextAllowedCityQueriesTests : BaseDatabaseTest
     [Fact]
     public async Task GetByCityAndStateAsync_ShouldBeCaseInsensitive()
     {
+        // Arrange
         var city = AllowedCityBuilder.AsTestCity("Muriaé", "MG").WithIbgeCode(3143906).Build();
         _context.AllowedCities.Add(city);
         await _context.SaveChangesAsync();
 
+        // Act
         var result1 = await _queries.GetByCityAndStateAsync("MURIAÉ", "mg");
         var result2 = await _queries.GetByCityAndStateAsync("muriaé", "MG");
 
+        // Assert
         result1.Should().NotBeNull();
         result1!.CityName.Should().Be("Muriaé");
         result2.Should().NotBeNull();
@@ -189,24 +225,36 @@ public class DbContextAllowedCityQueriesTests : BaseDatabaseTest
     [Fact]
     public async Task GetByCityAndStateAsync_WithNullCityName_ShouldReturnNull()
     {
+        // Arrange
+
+        // Act
         var result = await _queries.GetByCityAndStateAsync(null!, "MG");
 
+        // Assert
         result.Should().BeNull("normalização de null deve resultar em string vazia e não encontrar registro");
     }
 
     [Fact]
     public async Task IsCityAllowedAsync_WithNullCityName_ShouldReturnFalse()
     {
+        // Arrange
+
+        // Act
         var result = await _queries.IsCityAllowedAsync(null!, "MG");
 
+        // Assert
         result.Should().BeFalse("normalização de null deve resultar em string vazia");
     }
 
     [Fact]
     public async Task ExistsAsync_WithNullCityName_ShouldReturnFalse()
     {
+        // Arrange
+
+        // Act
         var result = await _queries.ExistsAsync(null!, "MG");
 
+        // Assert
         result.Should().BeFalse("normalização de null deve resultar em string vazia");
     }
 }
