+++ src/Modules/Locations/Tests/Unit/Application/ModuleApi/LocationsModuleApiTests.cs
@@ -1,11 +1,7 @@
-using FluentAssertions;
 using MeAjudaAi.Modules.Locations.Application.ModuleApi;
 using MeAjudaAi.Modules.Locations.Application.Services;
 using MeAjudaAi.Modules.Locations.Domain.ValueObjects;
 using MeAjudaAi.Shared.Geolocation;
-using Microsoft.Extensions.Logging;
-using Moq;
-using Xunit;
 
 namespace MeAjudaAi.Modules.Locations.Tests.Unit.Application.ModuleApi;
