+++ src/Modules/Locations/Tests/Unit/Infrastructure/Events/Handlers/AllowedCityDomainEventHandlersTests.cs
@@ -1,14 +1,7 @@
-using MeAjudaAi.Modules.Locations.Application.Queries;
-using MeAjudaAi.Modules.Locations.Domain.Entities;
 using MeAjudaAi.Modules.Locations.Domain.Events;
 using MeAjudaAi.Modules.Locations.Infrastructure.Events.Handlers;
-using MeAjudaAi.Shared.Events;
 using MeAjudaAi.Shared.Messaging;
 using MeAjudaAi.Shared.Messaging.Messages.Locations;
-using Microsoft.Extensions.Logging;
-using Moq;
-using Xunit;
-using FluentAssertions;
 
 namespace MeAjudaAi.Modules.Locations.Tests.Unit.Infrastructure.Events.Handlers;
 
@@ -22,11 +15,14 @@ public class AllowedCityDomainEventHandlersTests
     [Fact]
     public async Task CreatedHandler_HandleAsync_ShouldPublishIntegrationEvent()
     {
+        // Arrange
         var handler = new AllowedCityCreatedDomainEventHandler(_messageBusMock.Object, _loggerCreatedMock.Object);
         var domainEvent = new AllowedCityCreatedDomainEvent(Guid.NewGuid(), "City", "ST");
 
+        // Act
         await handler.HandleAsync(domainEvent, CancellationToken.None);
 
+        // Assert
         _messageBusMock.Verify(m => m.PublishAsync(
             It.Is<AllowedCityCreatedIntegrationEvent>(e => e.CityId == domainEvent.CityId && e.CityName == "City" && e.StateSigla == "ST"),
             null, It.IsAny<CancellationToken>()), Times.Once);
@@ -35,11 +31,14 @@ public class AllowedCityDomainEventHandlersTests
     [Fact]
     public async Task CreatedHandler_HandleAsync_ShouldLogInformation()
     {
+        // Arrange
         var handler = new AllowedCityCreatedDomainEventHandler(_messageBusMock.Object, _loggerCreatedMock.Object);
         var domainEvent = new AllowedCityCreatedDomainEvent(Guid.NewGuid(), "City", "ST");
 
+        // Act
         await handler.HandleAsync(domainEvent, CancellationToken.None);
 
+        // Assert
         _loggerCreatedMock.Verify(l => l.Log(
             LogLevel.Information,
             It.IsAny<EventId>(),
@@ -51,13 +50,16 @@ public class AllowedCityDomainEventHandlersTests
     [Fact]
     public async Task CreatedHandler_HandleAsync_WhenPublishFails_ShouldLogErrorAndRethrow()
     {
+        // Arrange
         var handler = new AllowedCityCreatedDomainEventHandler(_messageBusMock.Object, _loggerCreatedMock.Object);
         var domainEvent = new AllowedCityCreatedDomainEvent(Guid.NewGuid(), "City", "ST");
         _messageBusMock.Setup(m => m.PublishAsync(It.IsAny<object>(), null, It.IsAny<CancellationToken>()))
             .ThrowsAsync(new Exception("Pub fail"));
 
+        // Act
         await Assert.ThrowsAsync<Exception>(() => handler.HandleAsync(domainEvent, CancellationToken.None));
         
+        // Assert
         _loggerCreatedMock.Verify(l => l.Log(
             LogLevel.Error,
             It.IsAny<EventId>(),
@@ -69,11 +71,14 @@ public class AllowedCityDomainEventHandlersTests
     [Fact]
     public async Task UpdatedHandler_HandleAsync_ShouldPublishIntegrationEvent()
     {
+        // Arrange
         var handler = new AllowedCityUpdatedDomainEventHandler(_messageBusMock.Object, _loggerUpdatedMock.Object);
         var domainEvent = new AllowedCityUpdatedDomainEvent(Guid.NewGuid(), "NewCity", "NS");
 
+        // Act
         await handler.HandleAsync(domainEvent, CancellationToken.None);
 
+        // Assert
         _messageBusMock.Verify(m => m.PublishAsync(
             It.Is<AllowedCityUpdatedIntegrationEvent>(e => e.CityId == domainEvent.CityId && e.CityName == "NewCity" && e.StateSigla == "NS"),
             null, It.IsAny<CancellationToken>()), Times.Once);
@@ -82,11 +87,14 @@ public class AllowedCityDomainEventHandlersTests
     [Fact]
     public async Task DeletedHandler_HandleAsync_ShouldPublishIntegrationEvent()
     {
+        // Arrange
         var handler = new AllowedCityDeletedDomainEventHandler(_messageBusMock.Object, _loggerDeletedMock.Object);
         var domainEvent = new AllowedCityDeletedDomainEvent(Guid.NewGuid());
 
+        // Act
         await handler.HandleAsync(domainEvent, CancellationToken.None);
 
+        // Assert
         _messageBusMock.Verify(m => m.PublishAsync(
             It.Is<AllowedCityDeletedIntegrationEvent>(e => e.CityId == domainEvent.CityId),
             null, It.IsAny<CancellationToken>()), Times.Once);
