+++ src/Modules/Locations/Tests/Unit/Infrastructure/ExternalApis/ViaCepClientTests.cs
@@ -1,11 +1,9 @@
-using System.Net;
-using FluentAssertions;
 using MeAjudaAi.Modules.Locations.Domain.ValueObjects;
 using MeAjudaAi.Modules.Locations.Infrastructure.ExternalApis.Clients;
 using MeAjudaAi.Modules.Locations.Infrastructure.ExternalApis.Responses;
 using MeAjudaAi.Shared.Serialization;
 using Microsoft.Extensions.Logging.Abstractions;
-using Xunit;
+using System.Net;
 
 namespace MeAjudaAi.Modules.Locations.Tests.Unit.Infrastructure.ExternalApis;
 
@@ -18,7 +16,7 @@ public sealed class ViaCepClientTests : IDisposable
     public ViaCepClientTests()
     {
         _mockHandler = new MockHttpMessageHandler();
-        _httpClient = new HttpClient(_mockHandler)
+        _httpClient = new HttpClient(_mockHandler.GetHandler())
         {
             BaseAddress = new Uri("https://viacep.com.br/")
         };
@@ -163,52 +161,4 @@ public sealed class ViaCepClientTests : IDisposable
         _mockHandler.LastRequestUri.Should().Contain("ws/01001000/json/");
     }
 
-    // Mock HttpMessageHandler
-    private sealed class MockHttpMessageHandler : HttpMessageHandler, IDisposable
-    {
-        private HttpResponseMessage? _responseMessage;
-        private Exception? _exception;
-        public string? LastRequestUri { get; private set; }
-
-        public void SetResponse(HttpStatusCode statusCode, string content)
-        {
-            _responseMessage?.Dispose();
-            _responseMessage = new HttpResponseMessage(statusCode)
-            {
-                Content = new StringContent(content, System.Text.Encoding.UTF8, "application/json")
-            };
-            _exception = null;
-        }
-
-        public void SetException(Exception exception)
-        {
-            _responseMessage?.Dispose();
-            _exception = exception;
-            _responseMessage = null;
-        }
-
-        protected override Task<HttpResponseMessage> SendAsync(
-            HttpRequestMessage request,
-            CancellationToken cancellationToken)
-        {
-            LastRequestUri = request.RequestUri?.ToString();
-
-            if (_exception is not null)
-            {
-                throw _exception;
-            }
-
-            return Task.FromResult(_responseMessage ?? new HttpResponseMessage(HttpStatusCode.InternalServerError));
-        }
-
-        protected override void Dispose(bool disposing)
-        {
-            if (disposing)
-            {
-                _responseMessage?.Dispose();
-                _responseMessage = null;
-            }
-            base.Dispose(disposing);
-        }
-    }
 }
