+++ tests/MeAjudaAi.Shared.Tests/TestInfrastructure/Mocks/Http/MockHttpMessageHandler.cs
@@ -1,11 +1,12 @@
-using System.Net;
 using Moq.Protected;
+using System.Net;
 
 namespace MeAjudaAi.Shared.Tests.TestInfrastructure.Mocks.Http;
 
 /// <summary>
 /// Mock configurável de HttpMessageHandler para simular respostas HTTP em testes.
 /// Reutilizável por todos os módulos que fazem chamadas HTTP externas.
+/// Suporta tanto API simples (SetResponse/SetException) quanto API fluente (SetupResponse/SetupException).
 /// </summary>
 public sealed class MockHttpMessageHandler
 {
@@ -15,6 +16,63 @@ public sealed class MockHttpMessageHandler
     {
         _mockHandler = new Mock<HttpMessageHandler>();
     }
+
+    /// <summary>
+    /// URL da última requisição recebida pelo mock.
+    /// </summary>
+    public string? LastRequestUri { get; private set; }
+
+    #region API Simples (compatível com testes que usam SetResponse/SetException)
+
+    /// <summary>
+    /// Configura uma resposta padrão para qualquer requisição.
+    /// </summary>
+    public void SetResponse(HttpStatusCode statusCode, string content)
+    {
+        _mockHandler.Reset();
+        LastRequestUri = null;
+
+        _mockHandler
+            .Protected()
+            .Setup<Task<HttpResponseMessage>>(
+                "SendAsync",
+                ItExpr.IsAny<HttpRequestMessage>(),
+                ItExpr.IsAny<CancellationToken>())
+            .Callback<HttpRequestMessage, CancellationToken>((req, _) =>
+            {
+                LastRequestUri = req.RequestUri?.ToString();
+            })
+            .ReturnsAsync(new HttpResponseMessage
+            {
+                StatusCode = statusCode,
+                Content = new StringContent(content, System.Text.Encoding.UTF8, "application/json")
+            });
+    }
+
+    /// <summary>
+    /// Configura uma exceção padrão para qualquer requisição.
+    /// </summary>
+    public void SetException(Exception exception)
+    {
+        _mockHandler.Reset();
+        LastRequestUri = null;
+
+        _mockHandler
+            .Protected()
+            .Setup<Task<HttpResponseMessage>>(
+                "SendAsync",
+                ItExpr.IsAny<HttpRequestMessage>(),
+                ItExpr.IsAny<CancellationToken>())
+            .Callback<HttpRequestMessage, CancellationToken>((req, _) =>
+            {
+                LastRequestUri = req.RequestUri?.ToString();
+            })
+            .ThrowsAsync(exception);
+    }
+
+    #endregion
+
+    #region API Fluente (para cenários com múltiplos padrões de URL)
 
     /// <summary>
     /// Configura uma resposta para um padrão de URL específico.
@@ -36,7 +94,11 @@ public sealed class MockHttpMessageHandler
                     req.RequestUri != null &&
                     req.RequestUri.ToString().Contains(urlPattern)),
                 ItExpr.IsAny<CancellationToken>())
-            .ReturnsAsync(() => new HttpResponseMessage
+            .Callback<HttpRequestMessage, CancellationToken>((req, _) =>
+            {
+                LastRequestUri = req.RequestUri?.ToString();
+            })
+            .ReturnsAsync(new HttpResponseMessage
             {
                 StatusCode = statusCode,
                 Content = new StringContent(content, System.Text.Encoding.UTF8, contentType)
@@ -63,6 +125,10 @@ public sealed class MockHttpMessageHandler
                     req.RequestUri != null &&
                     req.RequestUri.ToString().Contains(urlPattern)),
                 ItExpr.IsAny<CancellationToken>())
+            .Callback<HttpRequestMessage, CancellationToken>((req, _) =>
+            {
+                LastRequestUri = req.RequestUri?.ToString();
+            })
 #pragma warning disable CA2000 // HttpResponseMessage em mock é ownership do caller
             .ReturnsAsync(new HttpResponseMessage
 #pragma warning restore CA2000
@@ -88,6 +154,10 @@ public sealed class MockHttpMessageHandler
                     req.RequestUri != null &&
                     req.RequestUri.ToString().Contains(urlPattern)),
                 ItExpr.IsAny<CancellationToken>())
+            .Callback<HttpRequestMessage, CancellationToken>((req, _) =>
+            {
+                LastRequestUri = req.RequestUri?.ToString();
+            })
             .ThrowsAsync(exception);
 
         return this;
@@ -130,5 +200,8 @@ public sealed class MockHttpMessageHandler
     public void Reset()
     {
         _mockHandler.Reset();
+        LastRequestUri = null;
     }
+
+    #endregion
 }
